+++ app/components/HomeScreenList.tsx
@@ -1,6 +1,6 @@
 import { StyleSheet, Text, View,TouchableOpacity } from 'react-native'
 import React from 'react'
-import styles from '../Styles'
+import styles from '../../Styles'
 import { FlatList } from 'react-native-gesture-handler'
 
 export default function HomeScreenList({categories,onListPress}) {
