+++ app/screens/index.js
@@ -1,0 +1,1 @@
+export {default as HomeScreen} from './Home/homeScreen'
